First commit!
This is a simple Arduino project. This is also my first time making
something beyond "make a LED blink" type of thing.

I don't really know C++ and I don't really know microcontrollers so this
is kinda a trial-by-fire project.

It's an oviously simple circuit. It's obviously amatuer C++ code. But it
works! And it's, like, fine in that respect.